pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "secure-agents"
version = "0.1.0"
description = "Secure on-prem AI agent framework - modular agents, tools, and LLM providers with defense-in-depth security"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "imapclient>=3.0",
    "pdfplumber>=0.10",
    "python-docx>=1.0",
    "httpx>=0.27",
    "click>=8.0",
    "structlog>=24.0",
    "docker>=7.0",
    "keyring>=25.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "watchdog>=4.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
gemini = ["google-genai>=1.0"]
gmail-oauth = ["google-auth-oauthlib>=1.0"]
all = ["secure-agents[anthropic,openai,gemini,gmail-oauth]"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
secure-agents = "secure_agents.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/secure_agents"]
